+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Version 1.2
 New Features
 ------------
 * Added additional before_response_sent event that gets fired even if an action was not invoked
+* Added method to add multiple custom routes at once
 
 
 Version 1.1